+++ docs/changes.rst
@@ -5,6 +5,7 @@ Release history
   Forwarding rule creation, update, and deletion was broken
 - feat: unit tests for call queue forwarding and selective forwarding rule creation and deletion
 - fix: missing return type for :meth:`wxc_sdk.workspace_locations.WorkspaceLocationApi.update`
+- fix: make parameter location_id optional in :meth:`wxc_sdk.telephony.devices.TelephonyDevicesApi.available_members`
 
 1.13.0
 ------